requirement.txt
python-telegram-bot==13.7
mysql-connector-python==8.0.26